Obsolete the minute provision plugin
Provide a fluent transition to existing minute plugin users by
providing a `minute.obsolete` method with lower priority. The
internal plugin will be picked once the internal package is
installed on the system, no action from the user is needed. Warn
users that the obsolete plugin will be removed in the future.

Co-authored-by: Petr Šplíchal <psplicha@redhat.com>