Fix login not working for cloud images
Handle better problems encountered when logging in to the instance.
In case cloud image standard message, inform the user about it:

    Login to server failed. Please login as the user 'cloud-user'
    rather then the user 'root'.

The second part of the sentence is directly taken from the ssh reply.
In case `authorized_keys` forces a weird command, which breaking login,
report it to the user as:

    Failed to push workdir to the machine. This usually means login as
    'root' to the test machine does not work.

Resolves #601

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>
Co-authored-by: Petr Šplíchal <psplicha@redhat.com>