Propagate options to guests based on the step
Previously, the guest parent was permanently set to provision step,
resulting in unexpected behaviour when commands such as
tmt run -a prepare -ddd were used, the debug options were not propagated
to the guest and hence the output from commands run on the guest from
prepare step was not shown. Change this behaviour so that the guest parent
is based on the step that's currently using it.

Signed-off-by: František Nečas <fnecas@redhat.com>