Install `rsync` before guest.pull()
@mruprich reported that after running installability test the `tmt`
fails on missing rsync when pulling out artifacts.

After investiation, it turned out, installability for `rsync` removes
`rsync` from the test machine, but `tmt` still expects it to be around.

This tries to fix the problem and make sure `rsync` is still installed
on the machine before trying to pull out the artifacts.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>