Make multiline shell scripts fail on error
Previously, if an error occurred in a pipeline or in a line of a
multiline script that is not last, the test would pass and a potential
error could go unnoticed. Use `set -eo pipefail` by default for
multiline shell scripts to avoid this problem.

This can make some tests fail, however for these tests it is possible to
explicitly disable these options. On the other hand, it is crucial that
tests fail when they are expected to which may not be the case without
these options set.

Signed-off-by: František Nečas <fnecas@redhat.com>