Ensure environment files are within metadata tree
Instead of using current working directory (which does not have to
be the root of the metadata tree) explicitly require relative
paths from the root and check them accordingly. Handle '=' on the
right side of the dotenv line. Extend the test coverage a bit.
Rebase on the latest master and use the ruamel yaml parser.