Split Guest class into two, to separate SSH-capable guests
The base class for all guests was expecting some parameters that were
sueful only in the context of SSH-capable, "connectable" guest. For
example Podman-backed guests have no use for `guest` or `password` keys.
To express this hierarchy better, and to open it for more strict types
involved in various `data` packages passed between methods, this patch
splits `Guest` class into a base one and a new one, representing a guest
with SSH capability.

Guest implemenations are now based on the new `GuestSSH` class, with the
exception of `GuestContainer` which is derived from `Guest` since it
does not use SSH.