Refactor location and signature of Phase's go() method
The method has been defined in `Phase` class, making it shared among all
of its subclasses. Unfortunately, some subclasses accept more specific
arguments, which makes their methods incompatible with superclass
definition.

This is a violation of Liskov substitution principle which states that
objects of superclass can be replaced with objects of subclass without
breaking an application. Note that while the implementation looks so,
`tmt` code is not actually using this freedom - not all subclasses of
`Phase` can be, and are not, in fact, used everywhere. For example,
`execute` plugins never meet `discovery` step.

The patch removes `go()` from `Phase`, and introduces more specific
methods into its subclasses. This involves adding two plugin-like base
classes replacing the original `Plugin`, because plugins that need to
be aware of a guest to do their job are common, therefore we now have a
base plugins class plus two thin wrappers with `go()` methods tailored
for their line of work.