Normalize step data to be always stored a list internally
`Step.data` was annotated as `Union[StepData, List[StepData]]`, allowing
both a single item and list of items. This complicates things when one
wants to do stuff like `self.data.get('foo')` - not possible with the
list. And if done in loop, one must first check `self.data` is actually
a list.

Which it was, the code in `Step.__init__` took care of that, but
annotations were puzzling. This patch settles down on `List[StepData]`
only, and changes normalization code to take advantage of this fact.