Polarion export fix component upload bug and upload id first
Component field can be both single element and list, depending on
project used, so the upload failed during export due to that.

Also moved the id generation to the start of export to Polarion and
added specific upload call for the id so the case export can be
continued more easily in case of errors.

Also fix a small bug in search by ID where search is not specific
enough.