Add serialization callbacks to data class fields
This greately reduces the need for custom `to/from_serialized` methods.
Should a field needed a custom serialization handling, such a code would
be "burried" in a methods taking care of serialization of the whole
class - this patch allows attaching the related code to fields directly,
making it much more local and visible. If the field requires special
handling, the implementation is attached to it right there.