Implement `deprecated` for obsoleted options
Add a new class `Deprecated` and a new parameter `deprecated` to
`tmt.utils.field()` to allow marking obsolete options with the version
since when it should not be used and an optional hint with the
recommended alternative.

For example the following field:

    tmt.utils.field(
        option="--old-option"
        deprecated=Deprecated(
            since="1.23",
            hint="Use --something-else instead."),
        help="I'm doing this and that.")

will produce this help message:

    --old-option   I'm doing this and that. The option is
                   deprecated since 1.23. Use --something-else
                   instead.

Also document the `multiple` parameter in the `field()` docstring.