Change `Plugin.go()` to accept logger and extra environment
This is part of the multihost support effort. Since a single phase can
be executed on multiple hosts at the same time, the `go()` method may
not rely on `self` when it comes to mutable state. Phase configuration -
fmf nodes, CLI options, step data - should be fine to *read*, but each
call to `go()` might be happening within different context, with a
different guest and logging expectations. Therefore the method must
become pure-ish, and modify data given to it.

There are probably more instances of this problem, but logging is the
first one we hit: to tell which output comes from which incarnation of a
single phase, whether it's the one running on guest A or the one on
guest B, `self.{info,verbose,...}` are useless, and `go()` must be given
a logger. The logger inside `Guest` instance given to `go()` *might* be
good enough, at the first glance, but I expect it to prove to be not
the case, therefore adding logger as new parameter.

And we already know we need to propagate extra info to tests and other
code, at least info about the guest they run on and its peers, therefore
we need a way to pass different set of envvars to invocations of `go()`.
So throwing `environment` as a parameter into this patch, because it's
"just" another parameter added to `go()`, and if we already know it's
needed, let's break the API once right away.