Support logging "topics" to allow lower unnecessary verbosity
Many verbose and debug messages are not interesting until a very
specific kind of error appears. This is especially true for key
normalization or command timestamps - both were added recently, both
increased verbosity a lot. This patch allows splitting debug and verbose
messages into "topics" that can be enabled when needed, e.g. when
reproducing an issue or debugging a user report.

```shell
$ tmt -vvdd run ...
$ tmt -vvdd --log-topic key-normalization run ...
```