Formalizing guest "facts" storage
I've been looking into tmt usage on Ubuntu/Debian, and while playing
with changes that might be needed to make things work - mostly
rpm-centric bits need tweaks - I slowly realized I'm adding bits that
rely on awareness of distro or distro family. This slowly resulted into
a refactoring of extra information tmt tracks about guests, besides the
IP address/hostname & keys defined by specification.

So, here are "facts", attached to each `Guest` instance, saved into
`guests.yaml`, documented, with primitives for easier adding new
fields. There is a slight increase in the overhead, with one extra
`Guest.execute()` call, but the previous worst cases should improve
because of a bit of caching. And I plan to improve the caching in the
future, but getting annotations right for cache, dataclass, property and
Python 3.6 is not as easy as I'd like.

```
default-0:
    role:
    guest: tmt-313-rNVYCvFt
    facts:
        in_sync: true
        distro: CentOS Stream 9
        kernel_release: 5.10.102.1-microsoft-standard-WSL2
        package_manager: dnf
        os_release_content:
            NAME: CentOS Stream
            VERSION: '9'
            ID: centos
            ID_LIKE: rhel fedora
            VERSION_ID: '9'
            PLATFORM_ID: platform:el9
            PRETTY_NAME: CentOS Stream 9
            ANSI_COLOR: 0;31
            LOGO: fedora-logo-icon
            CPE_NAME: cpe:/o:centos:centos:9
            HOME_URL: https://centos.org/
            BUG_REPORT_URL: https://bugzilla.redhat.com/
            REDHAT_SUPPORT_PRODUCT: Red Hat Enterprise Linux 9
            REDHAT_SUPPORT_PRODUCT_VERSION: CentOS Stream
        lsb_release_content: {}
        __class__:
            module: tmt.steps.provision
            name: GuestFacts
    image: centos:stream9
    user: root
    force_pull: false
    container: tmt-313-rNVYCvFt
    __class__:
        module: tmt.steps.provision.podman
        name: PodmanGuestData
```