Adds "bootstrap logger" for logging before CLI options are recognized (#1839)
Aims at parts of tmt lifetime where logging is needed, but CLI options
like --debug are unknown yet. This would be code that's executed in
import time, the most visible one is plugin exploration & importing, but
there may be others. At this moment, `fmf.utils.Logging` is used for this
purpose, but that's tmt's logging "sneaking" away from what `tmt.log`
offers and implements. And that's not good, as it may mean slightly
different logging behavior than the one exhibited during the rest of
tmt run.