Extract "run a command" functionality into a stand-alone helper
There are other places where running a command is a necessary tool, but
as they lay outside of `Common` class tree, they cannot use
`Common.run()`. To avoid duplication, and present "one obvious way to do
'it'", patch extracts the functionality into a stand-alone function, and
teaches `Common` class to use it.