Improve names and docs around CLI context in Common classes
A CLI context is saved, both as a class and instance members, and it's
used as a source of CLI options and fmf context. Unfortunatelly, the
word "context" is slightly overloaded, the patch tries to improve this
plus adding a couple of comments and more verbose docstrings.

* `Common._context` and `Common._options` attributes gained `_cli`
  prefix. This makes it clear what kind of context it is, and puts them
on par with already present `_fmf_context`.
* related methods have been renamed to match the added "cli" tag.
* also `context` parameter of `Common.__init__` is now called
`cli_context`.
* since `_cli_context` and `_cli_options` are variables, not methods,
methods serving the context or its parts are now properties. No more
`self._context` but `self._fmf_context()`.