Unify results examples in test and plan specification
The main "source of truth" when it comes to how results are saved,
should be `/spec/plans/results.fmf`. But, it seems useful to mention some
important bits and examples also in `/spec/tests/result.fmf`, especially
with regard to custom results produced by a test. This means both places
describe the same concept, including detailed examples, and it makes me
sad to find the almost same but slightly different examples in two
places like this. A pet peeve of mine.

So, the patch tries to unify the examples, adds missing bits to
`results.fmf` to make it complete, and, finally, adds a couple of
comments for future editors to treat `results.fmf` as the primary source.
Sharing examples between stories in an automated way does not seem to
be doable :(