Identify incorrect subcommand after a correct one
If a correct subcommand is followed by an incorrect one, the CLI should
raise a proper error. For example,
    $ tmt run --rm pl disc provin -h local
`disc` is the subcommand `discover` for short, but `provin` is not
`provision` for short, which is invalid actually. Hence, CLI should
not ignore the invalid short name of the subcommand `provision`.

Signed-off-by: Vector Li <huanli@redhat.com>
Co-authored-by: Miloš Prchlík <mprchlik@redhat.com>