Fix ruamel.yaml type waivers that mypy sometimes ignores
`ruamel.yaml` available to use does not set annotations of attributes
fully, which leaves mypy with autodetection. Infered types are the in
conflict with what we wish to assign to `YAML.width` and
`YAML.explicit_start`, even though our values are perfectly valid. The
first remedy, `type: ignore[assignment]` *used to* work - from time to
time, probably because of `ruamel.yaml` updates, mypy realizes we were
right and `ignore` is not needed, but this leads to failures when
running `pre-commit` locally, and it's far from being deterministic.

Therefore, giving up. Instead of `ignore`, use `cast()`, and change type
of our values - for mypy, that is - to match the attributes, being fully
aware they are incorrect. This should silence mypy until ruamel.yaml
gets fixed.