A custom wrapper for options instead of click.option()
Q: Ok, but... But... But why?!?

It might look as a pointless exercise or "not invented here" syndrome,
but hear me out :)

The patch adds a very trivial wrapper, `tmt.options.option`, which
accepts a bunch of parameters, and calls `click.option()` with them.
Effectively a `click.option` replacement. The added value?

* type annotations. The original `click.option` is annotated, but in the
very basic form, similar to `*args: Any, **kwargs: Any`. This prevents
no accidental mistakes like `is_flag=(1, 2)`.
* *stricter* type annotations. `click.option` has some functionality tmt
is not using, to accommodate said functionality, some parameters are
allowed the range of types wider than what can appear in tmt code. Our
wrapper gives us a chance to be stricter.
* `deprecated` support. A container describing a deprecated field, or an
option, but it's part of `tmt.utils.field` API - that's good - but
`click.option` does not accept it, and that's sad. Plenty of options
defined in `tmt.cli` are deprecated, wouldn't it be nice if we could use
the same approach to tracking deprecation we established for plugin
fields? Our `option()` accepts `Deprecated` container, adds it to help,
just like the `field()` does. Actually, since our `field()` calls our
`option()`, `field()` no longer has to care about rendering of
`deprecated` flag, and `option()` does that for both fields and options.

Not so pointless anymore, is it? :) In the future, we can teach our
`option()` few more tricks, like default `metavar` for `Choice` options,
or some custom defaults that would have to be repeated in every
`click.option()` call.