Cache resolved linters
Improves the speed of linting of multiple items: `Lintable.lint()` can
find out what linters to run, by combining lists of enabled and disabled
checks. The final list is not saved, but it also does not depend on the
actual object to be checked. Therefore the patch adds a bit of caching:
list of linters is created and re-used for checking object of the same
class (test/plan/story).

Simple change, but fairly measurable: linting all tests, plans and
stories takes around 4.5 seconds on my workstation, compared to 5.5
seconds before the patch.