Parallelize prepare and execute steps
This pull request implements support for parallel execution of the
`prepare` and `execute` step phases. This allows for the actual
multihost test support.

Both steps no longer process phases in sequence, instead a queue with
thread pool is at their core, and phases are "paired" together with
guests they should run on according to the `name`, `role` and `where`
fields. After this scheduling, phases are processed in sequence, with
phases bound to multiple guests executed on all their guests at the same
time.

The patch affects many aspects of tmt internals related to multihost
testing, from the scheduling mentioned above to logging and test
environment.