Make "needs sudo" a guest fact
Then it can be used by interested parties. It was needed for package
installation, then for installation of tmt scripts, and we seem to need
it for touching `/etc/hosts` in multi-host lands.

The patch adds the fact, but does not seek out all uses of `sudo` - some
will need a bit more work, therefore worth their own patches.