Make tmt test wrapper name unique to avoid race conditions
A single execute plugin may be running the same test in the context of
different guests at the same time. Using the same name for all of them
opens the door to race conditions, and the name must be unique - the
script becomes a shared mutable state, and it must be protected. We can
add locks, but creating a dedicated file costs us nothing, and it's easy
to tell to which branch of execution the file belongs to.