Test data path must use safe guest/test names
Guest and test names are part of test data path, but instead of raw
names, tmt must use the sanitized, "safe" variants. Only these are
treated and suitable to be path components.

As a side-effect, name/safe_name implementation has changed to use
`@property` to re-do `safe_name` when `name` changes. The original
attempt fixing the issue used `safe_name` - but it was outdated, because
some plugins modify a test `name` property without changing `safe_name`
as well, therefore changing the implementation so it would not be
possible to change one but not the other.