Add `pyright` as a `pre-commit` check
Another pair of eyes, pyright is a type check plus a bit more, and will
inspect the code from slightly different point of view. As a result, it
will report issues mypy cannot spot, or cannot spot *yet* because the
set of implemented checks is simply different.

The slow, gradual approach is the key here, like the one we know from
mypy introduction, and slowly files and packages will get covered.