Save click context in click context object
Click uses a `click.Context` instance to carry information needed for
managing commands. Apps may attach "context object" to it, in our case,
it's a `tmt.click.ContextObject`, and it's then reachable via
`click.Context.obj` property. But when looking at `ContextObject`, there
is not way to reach the parent `click.Context` - this patch fixes this.