Cleanup of "logging function" types
These were defined in the early days of adding annotations to
`tmt/utils.py`, but since the conversion to `tmt.log.Logger`, the
logging situation cleaned up nicely. Today, we use only a single type of
logging callbacks, reduce annotations and custom types and reflect the
simple times we live in in the code.