Replace self.opt() when looking for debug/verbose/quiet setting
`self.opt()` is used to acquire values of options, but it's hard to
annotate and ensure returned values are used correctly. The patch is
part of the series replaces type-unaware dispatcher methods with more
formal ways. The move of responsibility for logging from `Common` to
`Logger` together with dataclasses, step data and other makes `opt()`
much less needed, and it can be slowly replaced for the well-known and
shared options.