Expose "key address" to normalization callbacks (#1869)
Keys belong to various objects like step phases or fmf ids, and tmt can
do a better job when reporting unexpected input values. Another step is to
expose the location of the key to normalization callbacks and a
dedicated exception for reporting the violations in a unified manner.

The key address is already used by schema validation when reporting
validation errors, and while it's often subpar, it enhances error
messages a bit.

In the follow-up patches, improving the addresses will happen - for
example, sometimes a class name is printed instead of an fmf node name,
usually, because the node name is not available.