Use correct url for the release archive, fix docs
Fix the `Source0` url in the spec file so that it points to a
correct location from where `packit` can download the archive when
proposing downstream pull requests.

The `:ref:` text role cannot be used in the README as it would not
be correctly rendered in the man page and the long description on
the pypi server. Let's just describe the section location instead.
Documentation test updated to prevent such errors in the future.