Revert the `Source0` url to the original value
As for now we need to manually upload the source tarball to the
github release because the archive prepared by github does not
contain the generated man page and version is not substituted.

This also fixes the broken `make rpm` target. We should simplify
the tarball creation but that will be work for another patch.