Run all available tests only upon a user request
It does not make sense to run all enabled tests and plans for each
and every pull request change. Let's rather execute a core set of
tests for each update to ensure basic sanity of the change and run
the full set of available tests upon a user request using the
`/packit test --identifier full` comment.