Spec-based container becomes generic over input/output specs
Since the container class is often best aware of what type of
specification it accepts, and what type of specification it converts to,
it is trivial to make the class generic and allow derived classes to
inject the right types.

This should reduce the number of overriden methods where the only goal
was to provide more fitting parameter/return types.