Do not export fmf id's ref when it's the default branch
This has been sort of a thing already, some tools depend on `ref` not
being exported when it's "default branch", but after recent export and
git-related changes, `ref` started to pop out in exports.

Patch saves the "context" - git root, fmf root, default branch - for
each fmf id tmt creates. `ref` is now always set, and the decission
whether or not to display it is left to code that exposes the ids:
export, `show()` methods, etc. `tmt.utils.fmf_id()` is too low-level to
see the bigger picture and understand the context.