Add cache_property for things that are generated but not often
There is an existing support in recent Python versions, but the quality
of annotations varies because of RPM packaging different versions of
typing extensions. Patch adds a very basic `cached_property` with
correct annotations to be good enough for our use cases.

Also some properties have turned to this new decorator, to demonstrate
the use, but it's definitely the complete transition.