Wrap `click.Choice` use with `choices` parameter
Instead of exposing the internals, provide a nice `choices` parameter
accepting a list of choices. This decouple the semantics of "having a
set of choices" from the CLI-related implementation. It's a bit strange
in the CLI area of our code, but this is a general trend, already
established by `tmt.utils.field()`, and `option()` should follow the
same approach to how field/option choices are specified.