Render default friendly command for guest execution
While it's still possible for a plugin to provide its own friendly
command label, the defautl one needs to be rendered on the level of
`Guest.execute`. Leaving it for deeper methods, we risk the friendly
command to include SSH and environment details, making it too verbose
for user-friendly info.