Extend tmt-reboot to allow reboot from outside of the test process
Patch introduced a "test pidfile" in which the execute plugin would
store test process pid. This file is then used by `tmt-reboot` to
discover which process to signal when reboot is requested.

This allows `tmt-reboot` to be called from outside of the test process
tree while keeping tmt in the loop and correctly notified.