Add `skip` as a supported custom result outcome
The use case is when a single tmt test produces multiple results
but some of them were not executed because the process decided so.
It should be visible to the user that such a test exists and that
there is not a bug in the implementation which caused the test to
be missing.

Co-authored-by: Petr Šplíchal <psplicha@redhat.com>