Include more dependencies in the dev environment
Sometimes, one might need to run tmt tests against the development code,
therefore the dev environment shall work with testcloud or junit reporting.

We keep the test env in the game as it defines the test matrix - it's
based on dev environment, but one may use it to run tests against
well-defined Python version.