Use `UpdatableMessage` for execute/internal progress bar
Internal execute plugin was reporting plan progress, with no verbosity,
by updating a single line with test names as plan progresses. We already
do have a code for this, in `tmt.utils.updatable_message`, let's re-use
it.

The patch does a small refactoring to make the existing code more open,
and refactors the execute plugin to use a custom subclass of the
existing code, with extra touches the execute plugin needs.