Fix use of generics that are no longer "unsubscriptable"
Older Python forced us to ignore some generic types when being used as
base classes in our code. Older Python would not allow using subscript
needed for generic type parameter, and wrapping the type within a string
would not help as that would not be a base class then.

Now whe we dropped older Pythons, we can fix this gap.