Hide test/plan/story internal fields from export
Fields - class members of `Test`, `Plan` and `Story` - that are not
backed by tmt specification should not be part of the export - by
default, that is. We still need them to be saved in `tests.yaml` for tmt
to load them when necessary. Marking them as internal, and adding a
control parameter to export to expose them when requested.

Fixes https://github.com/teemtee/tmt/issues/2298