Convert test execution internals to use "invocation" bundle (#2469)
* Convert test execution internals to use "invocation" bundle

Execute plugins pass around test and guest as components of a test
invocation. But, there is no data structure or its instance that would
allow code to add more data to this invocation description, which seems
to be necessary for implementation of some checks. So, motivated by
this, the patch refactors code to use `TestInvocation` dataclass,
bundling together plugin, guest and test:

* As a by-product, we get fewer parameters being passed around.
  Invocation package is created once, and then passed around. There is
  no chance a call forgets to pass a guest where it's needed. Simpler
  method signatures, and it's clear what's required to run a test.

And for the original goal, this allows for check to store per-invocation
data: for the watchdog check (https://github.com/teemtee/tmt/issues/1523)
we will need a way for the check (running in its own thread) to notify
the main thread (running the test) that the guest has been marked as dead
and/or has been forcefully rebooted. Data like this cannot be added to
`Test` or `Guest` instances because those are shared - with the
`TestInvocation`, the check will be given this object, and would be
able to store whatever data it needs.

* squash: fix docstrings

* squash: address comments