Allow running upgrade from the current repository
The discover part of the `upgrade` execute step is just a slightly
modified version of the `discover` step which defaults to using
the current repository to search for available tests, in this case
for upgrade tasks. Let's allow having both real plans and upgrade
paths and tasks in the same repository for easier upgrade testing
development.