Move test-requested reboot handling into test invocation class
The reboot count is owned by a test invocation instance. All
reboot-related methods accept a single argument, the invocation
instance. This makes the code prime candidate for being owned by the
invocation class as well.

I added a bit of refactoring and polishing: the base is a mere move from
plugin-level methods into the `TestInvocation` class. On top of that, a
bit of reformatting added as I was fixing `self` and `invocation`
references, plus few recent goodies were used, namely the cached
property.

While working on a watchdog functionality, there will be an async
reboot request. As of now, the idea is for the watchdog to set a flag in
the test invocation instance, the flag which would be also consumed by
the `reboot_requested` property. Another point towards the invocations
owning reboot handling.