Custom soft/hard reboot commands for the connect provision plugin
To support a watchdog check
(https://github.com/teemtee/tmt/issues/1523), tmt must learn more about
rebooting a guest, especially when the SSH connection is no longer
available. The patch begins by extending the `connect` plugin to support
custom commands. This will help with the testing and local runs.

Imagine a Beaker box, reserved outside of tmt, used with `connect`
plugin. The test may render the box unresponsive, in which case the SSH
& a remote `reboot` command are not possible. For local runs and testing
we can provide `connect` with info to trigger reboot even in those
conditions, e.g. with `soft-reboot: bkr system-power --action reboot $IP`.

This comes with a bit of refactoring, to allow `reboot` to run a command
on the runner rather than on the guest.