docs: tweak docs to include how to see details about test failures
One common issue that most new users (like me) face when writing
tests is that the tests will not work flawlessly on the first try.

Hence seeing details what went wrong in the examples will help
to get started with tmt quicker. This commit adds examples for this
as small doc addition in the `Examples` section.

Signed-off-by: Michael Vogt <mvogt@redhat.com>