Generate plugin documentation from their sources
This required a small refactoring of how field properties were stored.
Instead of treating them as something owned by CLI options, they have
their use outside of CLI area, and must exist in field metadata to be
consumable in general.

Now we have a template, a script, a section in docs. Fixing the content
would be the next step...