AVC check now saves a timestamp on guest instead of using runner's time
Because of possible - and likely - difference in time between a tmt
runner and guests, AVC check may be asked to collect events that were
logged *before* the test started according to tmt's `start_time`
timestamp. This would obviously lead to AVC check collecting no events,
and reporting a success.

The patch instead saves a timestamp in a file on the guest, using proper
locale setting, and uses this timestamp instead of tmt's `start_time`.
This required a bit of refactoring, because now we need to run a shell
script *before* the test, and all our helpers and reporting was
implemented for the "after" phase only.

Co-authored-by: Petr Šplíchal <psplicha@redhat.com>
Co-authored-by: Lukáš Zachar <lzachar@redhat.com>