Introduce "topology" addresses for guests
Besides a current "guest" attribute - now called "primary address", the
patch adds "topology address" for guests to use when talking to each
other. As of now, the topology address of a guest is the same as its
primary address, but we expected that to change.

Related to https://github.com/teemtee/tmt/issues/2620